refactor mock_timer
